refactor: product projection 오류 수정
refactor: product projection 오류 수정

+++ product-api/src/main/java/org/teamsparta/productapi/domain/product/dto/request/ProductVariantRequest.java
@@ -1,10 +1,11 @@
 package org.teamsparta.productapi.domain.product.dto.request;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public record ProductVariantRequest(
         String sku,
-        Long price,
+        BigDecimal price,
         Integer stockQuantity,
         Map<String, Object> optionJson
 ) {
